games/pyproject.toml
[project]
name = "games"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.5",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
    "ty>=0.0.1a16",
]
